Create separate files. But test suite is removed. This one is easier for the eyes but I don't have a test suite anymore.
